refactor(rules-builder): extract DocumentPicker state to factory
